fix(DPLAN-12013): move addNewForm logic to the separate component
