+++ translations/messages+intl-icu.de.yml
@@ -1122,6 +1122,7 @@ email.text.notification_for_unsubmitted_statement.statements: >
     }
 end: "Ende"
 enabled: "freigeschaltet"
+entity.create: "{entity} anlegen"
 entity.administration.select.on.page: "Alle {entities} auf dieser Seite markieren"
 entities.marked.delete: "{sum} markierte {entities} löschen"
 entry: "Eingang"
@@ -3729,7 +3730,6 @@ tag.category: "Schlagwortkategorie"
 tag.category.name.unique.error: "Die Schlagwort-Kategorie existiert bereits. Bitte verwenden Sie einen anderen Namen."
 tag.category.new: "Neue Schlagwort-Kategorie"
 tag.category.new.create: "Schlagwort-Kategorie anlegen"
-tag.create: "Tag erstellen"
 tag.delete: "Schlagwort löschen"
 tag.move.toTopic: "in Thema verschieben"
 tag.move.toTopic.error: "Beim Schlagwort(e) Verschieben ist ein Fehler aufgetreten."
@@ -3742,7 +3742,7 @@ tag.name.empty.error: "Das Schlagwort muss einen Namen haben."
 tag.topic.name.empty.error: "Das Thema muss einen Namen haben."
 tag.topic.name.duplicate.error: "Das Thema existiert bereits. Bitte verwenden Sie einen anderen Namen."
 tag.new: "Neues Schlagwort"
-tag.new.create: "Schlagwort anlegen"
+tag.create.new: "Neues Schlagwort anlegen"
 tag.rename: "Schlagwort umbenennen"
 tag.search: "Schlagwort suchen"
 tag.topic.new: "Schlagwort/Thema anlegen…"
@@ -3917,10 +3917,10 @@ tool.active: "Aktives Werkzeug"
 topic: "Thema"
 topic.create.short: "Thema anlegen"
 topic.create.duplicated.title: "Das Thema existiert in diesem Verfahren bereits."
-topic.create: "Neues Thema anlegen"
+topic.create.new: "Neue Schlagwort-Kategorie anlegen"
 topic.delete: "Thema löschen"
 topic.insertFirstTag: "Ein Schlagwort hinzufügen."
-topic.insertTag: "Schlagwort zum Thema hinzufügen"
+topic.insertTag: "Schlagwort zur Schlagwort-Kategorie hinzufügen"
 topic.name: "Name des neuen Themas"
 # The 'topic.new' key works as a button label. The 3 dots intentionally differentiate it from select options nearby.
 topic.new: "Neues Thema…"
